+++ example/main.dart
@@ -1,6 +1,6 @@
 library color_chooser;
 
-import 'package:color_chooser/color_chooser.dart';
+import 'package:color_chooser/color_chooser.dart' show ColorChooserScreenMobile;
 import 'package:flutter/material.dart';
 
 void main() {
@@ -15,6 +15,7 @@ class ExampleApp extends StatelessWidget {
     return MaterialApp(
       home: ColorChooserScreenMobile(
         changeColorFunction: (c) {
+          // ignore: avoid_print
           print(c.toString());
         },
       ),