+++ lib/color_chooser.dart
@@ -3,5 +3,6 @@ library color_chooser;
 /// Import this File to have access to
 /// the different Color Chooser Screens.
 
+export 'src/color_chooser_screen.dart';
 export 'src/mobile/color_chooser_screen_mobile.dart';
 export 'src/desktop/color_chooser_screen_desktop.dart';